requirements.txt
# ------------------------------
# 核心功能依赖
# ------------------------------
# Telegram机器人核心库（稳定版，支持对话、命令处理、SOCKS代理）
python-telegram-bot[socks]==21.7
# HTTP请求库（处理与Misaka Danmaku API的交互）
requests==2.31.0
# 环境变量读取库（加载.env文件中的配置）
python-dotenv==1.0.0

# ------------------------------
# 日志与格式化依赖
# ------------------------------
# JSON格式日志（便于Docker日志收集与分析）
python-json-logger==2.0.7
# 日志颜色美化（开发环境可选，生产环境不影响）
colorlog==6.8.2

# ------------------------------
# 类型提示与代码健壮性
# ------------------------------
# 类型提示扩展（支持Python 3.8+的高级类型特性）
typing-extensions==4.12.2
# 数据验证（可选，若后续扩展API参数验证功能）
pydantic==2.8.2

# ------------------------------
# 安全与兼容性依赖
# ------------------------------
# SSL证书验证（确保HTTPS请求安全，修复部分环境证书问题）
certifi==2024.2.2
# 字符编码处理（避免中文等特殊字符乱码）
chardet==5.2.0
# URL解析（处理API请求中的URL参数）
urllib3==2.2.1

# ------------------------------
# 可选：开发调试依赖（生产环境可注释）
# ------------------------------
# 代码热重载（开发时修改代码无需重启机器人）
watchdog==3.0.0
# 调试工具（打印详细请求日志）
httpcore==1.0.5
